+++ src/us/kbase/shock/client/test/ShockTestCommon.java
@@ -18,8 +18,8 @@ import us.kbase.common.test.TestException;
 
 public class ShockTestCommon {
 	
-	public static final String SHOCKEXE = "test.shock.exe";
-	public static final String SHOCKVER = "test.shock.version";
+	public static final String BLOBEXE = "test.blobstore.exe";
+	public static final String MINIOEXE = "test.minio.exe";
 	public static final String MONGOEXE = "test.mongo.exe";
 	
 	public static final String TEST_TEMP_DIR = "test.temp.dir";
@@ -38,16 +38,16 @@ public class ShockTestCommon {
 		return getTestProperty(TEST_TEMP_DIR);
 	}
 	
+	public static String getMinioExe() {
+		return getTestProperty(MINIOEXE);
+	}
+	
+	public static String getBlobstoreExe() {
+		return getTestProperty(BLOBEXE);
+	}
+	
 	public static String getMongoExe() {
 		return getTestProperty(MONGOEXE);
-	}
-	
-	public static String getShockExe() {
-		return getTestProperty(SHOCKEXE);
-	}
-	
-	public static String getShockVersion() {
-		return getTestProperty(SHOCKVER);
 	}
 	
 	public static Path getJarsDir() {