+++ RELEASE_NOTES.md
@@ -1,3 +1,12 @@
+VERSION 0.2.0 (Released TDB)
+----------------------------
+
+BACKWARD INCOMPATIBILIES:
+
+- Changes to make the client fully compatible with the Blobstore
+  - getFormat() has been moved from ShockFileInformation to ShockNode.
+  - getVersion() has been removed from ShockNode.
+
 VERSION: 0.1.0 (Released 6/13/2019)
 -----------------------------------